requirements.txt
# Core framework and dependencies
Flask==3.1.3
Werkzeug==3.1.6
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
gunicorn==23.0.0

# LINE Bot SDK and requirements
line-bot-sdk==3.19.1
pydantic>=2.11.4
pydantic-core>=2.33.2

# MQTT client for garage door communication
paho-mqtt==2.1.0

# Security and secret management
google-cloud-secret-manager==2.25.0
google-cloud-storage>=2.0.0
google-cloud-datastore>=2.0.0
python-dotenv==1.1.1

# API security and limiting
Flask-Limiter==3.11.0

# Caching (Redis with fallback)
# Utilities
requests==2.32.5

# Testing dependencies
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
